refactor[next]: relocate shared fixture module out of ffront_tests
Move feature_tests/ffront_tests/ffront_test_utils.py to
integration_tests/cases_utils.py (sibling of cases.py) and update all 42
importers. The module defines the shared backend/mesh/dimension fixtures
(exec_alloc_descriptor, mesh_descriptor, grid descriptors, dimension/offset/
field aliases) used across unit/integration/regression tests and by cases.py
itself, so its home under the ffront feature tests was misleading and created
a top-level->deeply-nested dependency. cases.py and definitions.py keep their
current, already-appropriate locations.
