+++ tests/next_tests/integration_tests/feature_tests/dace_tests/test_program.py
@@ -12,7 +12,7 @@ from gt4py import next as gtx
 from gt4py.next import common
 
 from next_tests.integration_tests import cases
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     Cell,
     Edge,
     IDim,
