+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_bound_args.py
@@ -13,7 +13,7 @@ from gt4py.next import int32
 
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import cartesian_case
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
     mesh_descriptor,
 )
