+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_closure_vars.py
@@ -12,7 +12,7 @@ import gt4py.next as gtx
 
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import cartesian_case
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
