+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_compiled_program.py
@@ -30,7 +30,7 @@ from next_tests.integration_tests.cases import (
     mesh_descriptor,
     unstructured_case,
 )
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     MeshDescriptor,
     exec_alloc_descriptor,
     simple_mesh,
