+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_concat_where.py
@@ -13,7 +13,7 @@ from gt4py import next as gtx
 from gt4py.next import broadcast
 from gt4py.next.ffront.experimental import concat_where
 from next_tests.integration_tests import cases
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
