+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_decorator.py
@@ -18,7 +18,7 @@ from gt4py.next.iterator import ir as itir
 from gt4py.next import common as gtx_common
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import cartesian_case
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
     simple_cartesian_grid,
 )
