+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_domain.py
@@ -17,7 +17,7 @@ from next_tests.integration_tests.cases import (
     JDim,
     cartesian_case,
 )
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
