+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_external_local_field.py
@@ -14,7 +14,7 @@ from gt4py.next import common, int32, neighbor_sum, astype, int64
 
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import V2E, Edge, V2EDim, Vertex, unstructured_case
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
     mesh_descriptor,
 )
