+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_gt4py_builtins.py
@@ -35,7 +35,7 @@ from next_tests.integration_tests.cases import (
     unstructured_case,
     unstructured_case_3d,
 )
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
     mesh_descriptor,
 )
