+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_import_from_mod.py
@@ -16,7 +16,7 @@ from next_tests import integration_tests
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import cartesian_case, IDim, KDim
 
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
     mesh_descriptor,
 )
