+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_math_builtin_execution.py
@@ -26,7 +26,7 @@ from gt4py.next.type_system import type_translation
 
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import IDim, cartesian_case, unstructured_case
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
 from next_tests.integration_tests.feature_tests.math_builtin_test_data import math_builtin_test_data
