+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_named_collections.py
@@ -20,7 +20,7 @@ import dataclasses
 
 from next_tests.integration_tests import cases
 from next_tests.integration_tests.cases import IDim, JDim, KDim, cartesian_case
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
