+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_scan.py
@@ -21,7 +21,7 @@ from next_tests.integration_tests.cases import (
     Koff,
     cartesian_case,
 )
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
