+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_type_conversion.py
@@ -24,7 +24,7 @@ from next_tests.integration_tests.cases import (
     cartesian_case,
     unstructured_case,
 )
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
     mesh_descriptor,
 )
