+++ tests/next_tests/integration_tests/feature_tests/iterator_tests/test_extractors.py
@@ -15,7 +15,7 @@ import pytest
 from gt4py import next as gtx
 from gt4py.next.iterator.transforms import extractors
 
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     IDim,
     JDim,
     KDim,
