+++ tests/next_tests/regression_tests/ffront_tests/test_cse_extraction_scan_init.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from next_tests.integration_tests.cases import KDim, cartesian_case
 from next_tests.integration_tests import cases
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     exec_alloc_descriptor,
 )
