+++ tests/next_tests/unit_tests/iterator_tests/test_type_inference.py
@@ -43,7 +43,7 @@ from next_tests.integration_tests.cases import (
     mesh_descriptor,
     unstructured_case,
 )
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import simple_mesh
+from next_tests.integration_tests.cases_utils import simple_mesh
 
 bool_type = ts.ScalarType(kind=ts.ScalarKind.BOOL)
 int_type = ts.ScalarType(kind=ts.ScalarKind.INT32)
