+++ tests/next_tests/unit_tests/program_processor_tests/runners_tests/dace_tests/test_dace_domain.py
@@ -19,7 +19,7 @@ from gt4py.next.program_processors.runners.dace.lowering import (
 from gt4py.next.program_processors.runners.dace import sdfg_args as gtx_dace_args
 from gt4py.next.iterator.ir_utils import domain_utils, ir_makers as im
 
-from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
+from next_tests.integration_tests.cases_utils import (
     Cell,
     KDim,
     Vertex,
